+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Aliased `melt` to `unpivot` for `DataFrame` and `LazyFrame`
 - Changed signature of `pivot` and `melt` methods
 - Changed signature of `date_range` and `date_ranges` methods
+- Changed `set_sorted` method to only accept a single column
 - Removed `use_earliest` option from `replace_time_zone` and `to_datetime` methods (use `ambiguous` instead)
 - Removed `by` and `closed` options from `rolling_*` methods (use `rolling_*_by` instead)
 - Removed `explode` method from `StringExpr` (use `split("").explode` instead)